feat: Segregated working with cart into separate CartManager
