+++ personalization-sdk/src/main/kotlin/com/personalization/api/managers/CartManager.kt
@@ -1,0 +1,10 @@
+package com.personalization.api.managers
+
+import com.personalization.api.OnApiCallbackListener
+
+interface CartManager {
+
+    fun getClientShoppingCart(
+        listener: OnApiCallbackListener?
+    )
+}
