+++ personalization-sdk/src/main/kotlin/com/personalization/api/managers/ProductsManager.kt
@@ -19,8 +19,4 @@ interface ProductsManager {
         itemId: String,
         listener: OnApiCallbackListener?
     )
-
-    fun getClientShoppingCart(
-        listener: OnApiCallbackListener?
-    )
 }
