+++ personalization-sdk/src/main/kotlin/com/personalization/features/products/impl/ProductsManagerImpl.kt
@@ -49,14 +49,6 @@ internal class ProductsManagerImpl @Inject constructor(
         )
     }
 
-    override fun getClientShoppingCart(listener: OnApiCallbackListener?) {
-        sendNetworkMethodUseCase.getAsync(
-            method = GET_CLIENT_SHOPPING_CART,
-            params = Params().build(),
-            listener = listener
-        )
-    }
-
     private fun Params.buildParams(
         itemId: String? = null,
         brands: String? = null,
@@ -84,7 +76,6 @@ internal class ProductsManagerImpl @Inject constructor(
     companion object {
         const val GET_PRODUCT_LIST_REQUEST = "products"
         const val GET_PRODUCT_INFO_REQUEST = "products/get"
-        const val GET_CLIENT_SHOPPING_CART = "products/cart"
 
         private const val LIMIT_KEY = "limit"
         private const val PAGE_KEY = "page"
